requirements-aws.txt
# AWS Dependencies for cloud deployment
boto3>=1.26.137
aws-cdk-lib>=2.87.0
constructs>=10.0.0
aws-lambda-powertools>=2.20.0

# For Lambda deployment packaging
lambda-uploader>=1.2.0

# AWS CLI alternative in Python
awscli>=1.29.0
